finish header layout
