+++ src/app/styles/app.css
@@ -7,3 +7,8 @@
     min-height: 100vh;
     overflow: hidden;
 }
+
+a {
+    color: inherit;
+    text-decoration: none;
+}
